requirements.txt
pytorch-lightning
ray[tune]
optuna
wandb
deepspeed